+++ pipelines/validation.jl
@@ -1,6 +1,6 @@
 module Validation
 
-function validate_data(data, max_sequence_length)
+function validate_data(data, max_sequence_length::Int64)
     # Check that all sequences are the correct length
     for sequence in data
         if length(sequence) != max_sequence_length
